+++ vscode/src/chat/MessageProvider.ts
@@ -54,6 +54,13 @@ const SAFETY_PROMPT_TOKENS = 100
 const nonDisplayTopics = new Set(['fixup'])
 
 /**
+ * The types of errors that should be handled from MessageProvider.
+ * `transcript`: Errors that can be displayed directly within a chat transcript, if available.
+ * `system`: Errors that should be handled differently, e.g. alerted to the user.
+ */
+export type MessageErrorType = 'transcript' | 'system'
+
+/**
  * A derived class of MessageProvider must implement these handler methods.
  * This contract ensures that MessageProvider is focused solely on building, sending and receiving messages.
  * It does not assume anything about how those messages will be displayed to the user.
@@ -61,10 +68,9 @@ const nonDisplayTopics = new Set(['fixup'])
 abstract class MessageHandler {
     protected abstract handleTranscript(transcript: ChatMessage[], messageInProgress: boolean): void
     protected abstract handleHistory(history: UserLocalHistory): void
-    protected abstract handleError(errorMsg: string): void
     protected abstract handleSuggestions(suggestions: string[]): void
     protected abstract handleCodyCommands(prompts: [string, CodyPrompt][]): void
-    protected abstract handleTranscriptErrors(transciptError: boolean): void
+    protected abstract handleError(errorMsg: string, type: MessageErrorType): void
 }
 
 export interface MessageProviderOptions {
@@ -123,13 +129,16 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
         this.contextProvider.configurationChangeEvent.event(() => this.sendCodyCommands())
     }
 
-    protected async init(): Promise<void> {
+    protected async init(chatID?: string): Promise<void> {
         this.loadChatHistory()
         this.sendTranscript()
         this.sendHistory()
-        await this.loadRecentChat()
         await this.contextProvider.init()
         await this.sendCodyCommands()
+
+        if (chatID) {
+            await this.restoreSession(chatID)
+        }
     }
 
     public async clearAndRestartSession(): Promise<void> {
@@ -152,6 +161,7 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
         // Reset the current transcript
         this.transcript = new Transcript()
         await this.clearAndRestartSession()
+        this.sendHistory()
         telemetryService.log('CodyVSCodeExtension:clearChatHistoryButton:clicked', undefined, { hasV2Event: true })
         telemetryRecorder.recordEvent('cody.messageProvider.clearChatHistoryButton', 'clicked')
     }
@@ -270,7 +280,7 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
                     err = 'Cody could not respond due to network error.'
                 }
                 // Display error message as assistant response
-                this.transcript.addErrorAsAssistantResponse(err)
+                this.handleError(err, 'transcript')
                 // We ignore embeddings errors in this instance because we're already showing an
                 // error message and don't want to overwhelm the user.
                 void this.onCompletionEnd(true)
@@ -308,7 +318,7 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
 
     public async executeRecipe(recipeId: RecipeID, humanChatInput = '', source?: ChatEventSource): Promise<void> {
         if (this.isMessageInProgress) {
-            this.handleError('Cannot execute multiple recipes. Please wait for the current recipe to finish.')
+            this.handleError('Cannot execute multiple recipes. Please wait for the current recipe to finish.', 'system')
             return
         }
 
@@ -321,24 +331,32 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
         humanChatInput = command?.text
         recipeId = command?.recipeId
 
-        logDebug('ChatViewProvider:executeRecipe', recipeId, { verbose: humanChatInput })
+        logDebug('MessageProvider:executeRecipe', recipeId, { verbose: humanChatInput })
 
         const recipe = this.getRecipe(recipeId)
         if (!recipe) {
-            logDebug('ChatViewProvider:executeRecipe', 'no recipe found')
+            logDebug('MessageProvider:executeRecipe', 'no recipe found')
             return
         }
 
         // Create a new multiplexer to drop any old subscribers
         this.multiplexer = new BotResponseMultiplexer()
 
-        const interaction = await recipe.getInteraction(humanChatInput, {
-            editor: this.editor,
-            intentDetector: this.intentDetector,
-            codebaseContext: this.contextProvider.context,
-            responseMultiplexer: this.multiplexer,
-            firstInteraction: this.transcript.isEmpty,
-        })
+        let interaction: Interaction | null = null
+
+        try {
+            interaction = await recipe.getInteraction(humanChatInput, {
+                editor: this.editor,
+                intentDetector: this.intentDetector,
+                codebaseContext: this.contextProvider.context,
+                responseMultiplexer: this.multiplexer,
+                firstInteraction: this.transcript.isEmpty,
+            })
+        } catch (error: any) {
+            this.handleError(error.message, 'system')
+            return
+        }
+
         if (!interaction) {
             return
         }
@@ -609,7 +627,6 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
         })
         this.transcript.addInteraction(interaction || customInteraction)
         this.sendTranscript()
-        this.handleTranscriptErrors(true)
         await this.saveTranscriptToChatHistory()
     }
 
@@ -639,6 +656,7 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
      * Save chat history
      */
     private async saveChatHistory(): Promise<void> {
+        this.loadChatHistory()
         const userHistory = {
             chat: MessageProvider.chatHistory,
             input: MessageProvider.inputHistory,
@@ -694,23 +712,6 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
     }
 
     /**
-     * Loads the most recent chat
-     */
-    private async loadRecentChat(): Promise<void> {
-        const localHistory = localStorage.getChatHistory()
-        if (localHistory) {
-            const chats = localHistory.chat
-            const sortedChats = Object.entries(chats).sort(
-                (a, b) => +new Date(b[1].lastInteractionTimestamp) - +new Date(a[1].lastInteractionTimestamp)
-            )
-            const chatID = sortedChats[0][0]
-            if (chatID !== this.currentChatID) {
-                await this.restoreSession(chatID)
-            }
-        }
-    }
-
-    /**
      * Send history to view
      */
     private sendHistory(): void {
@@ -730,8 +731,7 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
         const searchErrors = this.contextProvider.context.getEmbeddingSearchErrors()
         // Display error message as assistant response for users with indexed codebase but getting search errors
         if (this.contextProvider.context.checkEmbeddingsConnection() && searchErrors) {
-            this.transcript.addErrorAsAssistantResponse(searchErrors)
-            this.handleTranscriptErrors(true)
+            this.handleError(searchErrors, 'transcript')
             logError('ChatViewProvider:onLogEmbeddingsErrors', '', { verbose: searchErrors })
         }
     }